+++ extensions/intellij-lam/src/main/kotlin/org/lam/intellij/highlight/LamSyntaxHighlighterFactory.kt
@@ -1,0 +1,12 @@
+package org.lam.intellij.highlight
+
+import com.intellij.openapi.fileTypes.SyntaxHighlighter
+import com.intellij.openapi.fileTypes.SyntaxHighlighterFactory
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.vfs.VirtualFile
+
+/** Фабрика подсветки Lam (регистрируется в `plugin.xml`, задача 0022-02). */
+class LamSyntaxHighlighterFactory : SyntaxHighlighterFactory() {
+    override fun getSyntaxHighlighter(project: Project?, virtualFile: VirtualFile?): SyntaxHighlighter =
+        LamSyntaxHighlighter()
+}